backend/requirements.txt
fastapi
uvicorn
pydantic
feedparser
pyyaml
requests
